Add Options to toggle_term() (#2)
This change adds an options argument to toggle_term that allows the
caller to specify whether the mark should be updated or not.

I use this in my config to prevent intervention from marking a netrw
buffer if I toggle to terminal from the file explorer since I have netrw
configured to wipe on hide.